BF: Create copy of stimulus
The stimulus which is to be selected is copied to self._stimulus. Before, we only created a reference, which caused problems when a trigger line was used, because it modifies self._stimulus. We don't want to change to _original_ stimulus in self._stimuli, though, so we copy the stimulus to self._stimulus now. Changes there won't influence self._stimuli anymore.